Move Easy settings into corresponding folder
