+++ client/extensions/wp-super-cache/components/easy/easy.jsx
